requirements.txt
requests>=2.28.0
PyYAML>=6.0
Jinja2>=3.1.0
Markdown>=3.4.0